GNG-492: Refactor Components page directory structure
